+++ packages/component/src/Middleware/GroupActivities/createCoreMiddleware.js
@@ -1,3 +1,9 @@
+import { Constants } from 'botframework-webchat-core';
+
+const {
+  ActivityClientState: { SEND_FAILED }
+} = Constants;
+
 function bin(items, grouping) {
   let lastGroup;
   const groups = [];
@@ -17,11 +23,19 @@ function bin(items, grouping) {
   return groups;
 }
 
+function sendFailed(activity) {
+  return activity.from.role === 'user' && activity.channelData && activity.channelData.state === SEND_FAILED;
+}
+
 function shouldGroupTimestamp(activityX, activityY, groupTimestamp) {
   if (groupTimestamp === false) {
     // Hide timestamp for all activities.
     return true;
   } else if (activityX && activityY) {
+    if (sendFailed(activityX) || sendFailed(activityY)) {
+      return false;
+    }
+
     groupTimestamp = typeof groupTimestamp === 'number' ? groupTimestamp : Infinity;
 
     const timeX = new Date(activityX.timestamp).getTime();